Wed, Feb 8, 2023, 10:45 PM +00:00

+++ cs50/readability/readability.c
@@ -16,8 +16,8 @@ int main(void)
     int words = count_words(all);
     int sent = count_sentences(all);
 
-    float L = ((letters/words) * 100);
-    float S = ((sent/words) * (int) 100);
+    float L = ((letters/(float)words) * 100);
+    float S = ((sent/(float)words) * (int) 100);
